+++ src/components/SingleMyPost/SingleMyPost.jsx
@@ -23,22 +23,25 @@ const SingleMyPost = ({
   const navigate = useNavigate();
 
   const deletePost = async () => {
-    setIsLoading(true);
-    try {
-      const response = await deleteData(
-        DELETE_MY_POST(boardType, postId),
-        {
-          Authorization: `${localStorage.getItem('grantType')} ${localStorage.getItem('AToken')}`,
-        },
-        {},
-      );
-      if (response.status === 200) {
-        setIsDeleted(true); // 삭제 상태 업데이트
+    const confirmDelete = window.confirm('해당 글을 삭제하시겠습니까?');
+    if (confirmDelete) {
+      setIsLoading(true);
+      try {
+        const response = await deleteData(
+          DELETE_MY_POST(boardType, postId),
+          {
+            Authorization: `${localStorage.getItem('grantType')} ${localStorage.getItem('AToken')}`,
+          },
+          {},
+        );
+        if (response.status === 200) {
+          setIsDeleted(true); // 삭제 상태 업데이트
+        }
+      } catch (error) {
+        setError(true);
+      } finally {
+        setIsLoading(false);
       }
-    } catch (error) {
-      setError(true);
-    } finally {
-      setIsLoading(false);
     }
   };
 
@@ -62,7 +65,14 @@ const SingleMyPost = ({
       <s.PostContainer>
         <s.Top>
           <s.Title>{title}</s.Title>
-          <s.Time>{formatTime}</s.Time>
+          <s.Delete
+            onClick={(e) => {
+              e.stopPropagation(); // 이벤트 버블링 방지
+              deletePost();
+            }}
+          >
+            삭제
+          </s.Delete>
         </s.Top>
 
         <s.ContentText>{content}</s.ContentText>
@@ -80,14 +90,6 @@ const SingleMyPost = ({
           <span style={{ color: '#92A5BC' }}>{comment}</span>
         </s.Info>
       </s.PostContainer>
-      <s.Delete
-        onClick={(e) => {
-          e.stopPropagation(); // 이벤트 버블링 방지
-          deletePost();
-        }}
-      >
-        삭제
-      </s.Delete>
     </s.PostWrapper>
   );
 };
